+++ tests/image/test_inception.py
@@ -124,3 +124,28 @@ def test_compare_is(tmpdir):
     tm_mean, _ = metric.compute()
 
     assert torch.allclose(tm_mean.cpu(), torch.tensor([torch_fid["inception_score_mean"]]), atol=1e-3)
+
+
+@pytest.mark.skipif(not torch.cuda.is_available(), reason="test is too slow without gpu")
+@pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
+def test_store_on_cpu_arg(tmpdir, feature=64):
+    """Check that the store_on_cpu arg correctly keeps the states on cpu."""
+    metric = InceptionScore(feature=feature, store_on_cpu=True).cuda()
+
+    # Generate some synthetic data
+    img1 = torch.randint(0, 180, (10, 3, 299, 299), dtype=torch.uint8)
+    img2 = torch.randint(100, 255, (10, 3, 299, 299), dtype=torch.uint8)
+
+    batch_size = 2
+    for i in range(img1.shape[0] // batch_size):
+        metric.update(img1[batch_size * i : batch_size * (i + 1)].cuda(), real=True)
+
+    for i in range(img2.shape[0] // batch_size):
+        metric.update(img2[batch_size * i : batch_size * (i + 1)].cuda(), real=False)
+
+    assert "cuda" in str(metric.device)
+    assert all(str(feature.device) == "cpu" for feature in metric.real_features)
+    assert all(str(feature.device) == "cpu" for feature in metric.fake_features)
+
+    val = metric.compute()
+    assert str(val.device) == "cpu"